+++ src/Kiota.Builder/Writers/Python/CodeMethodWriter.cs
@@ -8,7 +8,10 @@ using Kiota.Builder.Extensions;
 namespace Kiota.Builder.Writers.Python;
 public class CodeMethodWriter : BaseElementWriter<CodeMethod, PythonConventionService>
 {
-    public CodeMethodWriter(PythonConventionService conventionService) : base(conventionService){}
+    public CodeMethodWriter(PythonConventionService conventionService, bool usesBackingStore) : base(conventionService){
+        _usesBackingStore = usesBackingStore;
+    }
+    private readonly bool _usesBackingStore;
     public override void WriteCodeElement(CodeMethod codeElement, LanguageWriter writer)
     {
         if(codeElement == null) throw new ArgumentNullException(nameof(codeElement));
@@ -125,34 +128,42 @@ public class CodeMethodWriter : BaseElementWriter<CodeMethod, PythonConventionSe
             foreach(var module in serializationModules)
                 writer.WriteLine($"{methodName}({module})");
     }
+    private CodePropertyKind[] _DirectAccessProperties;
+    private CodePropertyKind[] DirectAccessProperties { get {
+        if(_DirectAccessProperties == null) {
+            var directAccessProperties = new List<CodePropertyKind> {
+                CodePropertyKind.BackingStore,
+                CodePropertyKind.RequestBuilder,
+                CodePropertyKind.UrlTemplate,
+                CodePropertyKind.PathParameters
+            };
+            if(!_usesBackingStore) {
+                directAccessProperties.Add(CodePropertyKind.AdditionalData);
+            }
+            _DirectAccessProperties = directAccessProperties.ToArray();
+        }
+        return _DirectAccessProperties;
+    }}
+    private CodePropertyKind[] _SetterAccessProperties;
+    private CodePropertyKind[] SetterAccessProperties {
+        get {
+            if (_SetterAccessProperties == null) {
+                _SetterAccessProperties = new CodePropertyKind[] {
+                    CodePropertyKind.AdditionalData, //additional data and custom properties need to use the accessors in case of backing store use
+                    CodePropertyKind.Custom
+                }.Except(DirectAccessProperties)
+                .ToArray();
+            }
+            return _SetterAccessProperties;
+        }
+    }
     private void WriteConstructorBody(CodeClass parentClass, CodeMethod currentMethod, LanguageWriter writer, bool inherits) {
         if(inherits)
             writer.WriteLine("super().__init__()");
-        var propertiesWithDefaultValues = new List<CodePropertyKind> {
-            CodePropertyKind.AdditionalData,
-            CodePropertyKind.BackingStore,
-            CodePropertyKind.RequestBuilder,
-            CodePropertyKind.UrlTemplate,
-            CodePropertyKind.PathParameters,
-        };
-        foreach (var propWithoutDefault in parentClass.Properties.Except(parentClass.GetPropertiesOfKind(propertiesWithDefaultValues.ToArray()))
-                                        .Except(parentClass.GetPropertiesOfKind(CodePropertyKind.RequestAdapter))
-                                        .OrderByDescending(x => x.Kind)
-                                        .ThenBy(x => x.Name)) {
-            var returnType = conventions.GetTypeString(propWithoutDefault.Type, propWithoutDefault, true, writer);
-            conventions.WriteInLineDescription(propWithoutDefault.Description, writer);
-            writer.WriteLine($"self.{conventions.GetAccessModifier(propWithoutDefault.Access)}{propWithoutDefault.NamePrefix}{propWithoutDefault.Name.ToSnakeCase()}: {(propWithoutDefault.Type.IsNullable ? "Optional[" : string.Empty)}{returnType}{(propWithoutDefault.Type.IsNullable ? "]" : string.Empty)} = None");
-            writer.WriteLine();
-        }
-        foreach(var propWithDefault in parentClass.GetPropertiesOfKind(propertiesWithDefaultValues.ToArray())
-                                        .Where(x => !string.IsNullOrEmpty(x.DefaultValue))
-                                        .OrderByDescending(x => x.Kind)
-                                        .ThenBy(x => x.Name)) {
-            var returnType = conventions.GetTypeString(propWithDefault.Type, propWithDefault, true, writer);
-            conventions.WriteInLineDescription(propWithDefault.Description, writer);
-            writer.WriteLine($"self.{conventions.GetAccessModifier(propWithDefault.Access)}{propWithDefault.NamePrefix}{propWithDefault.Name.ToSnakeCase()}: {(propWithDefault.Type.IsNullable ? "Optional[" : string.Empty)}{returnType}{(propWithDefault.Type.IsNullable ? "]" : string.Empty)} = {propWithDefault.DefaultValue}");
-            writer.WriteLine();
-        }
+        WriteDirectAccessProperties(parentClass, writer);
+        WriteSetterAccessProperties(parentClass, writer);
+        WriteSetterAccessPropertiesWithoutDefaults(parentClass, writer);
+
         if(parentClass.IsOfKind(CodeClassKind.RequestBuilder)) {
             if(currentMethod.IsOfKind(CodeMethodKind.Constructor) &&
             currentMethod.Parameters.FirstOrDefault(x => x.IsOfKind(CodeParameterKind.PathParameters)) is CodeParameter pathParametersParam) {
@@ -166,6 +177,35 @@ public class CodeMethodWriter : BaseElementWriter<CodeMethod, PythonConventionSe
                 AssignPropertyFromParameter(parentClass, currentMethod, CodeParameterKind.PathParameters, CodePropertyKind.PathParameters, writer, conventions.TempDictionaryVarName);
             }
             AssignPropertyFromParameter(parentClass, currentMethod, CodeParameterKind.RequestAdapter, CodePropertyKind.RequestAdapter, writer);
+        }
+    }
+    private void WriteDirectAccessProperties(CodeClass parentClass, LanguageWriter writer) {
+        foreach(var propWithDefault in parentClass.GetPropertiesOfKind(DirectAccessProperties)
+                                        .Where(static x => !string.IsNullOrEmpty(x.DefaultValue))
+                                        .OrderByDescending(static x => x.Kind)
+                                        .ThenBy(static x => x.Name)) {
+            var returnType = conventions.GetTypeString(propWithDefault.Type, propWithDefault, true, writer);
+            conventions.WriteInLineDescription(propWithDefault.Description, writer);
+            writer.WriteLine($"self.{conventions.GetAccessModifier(propWithDefault.Access)}{propWithDefault.NamePrefix}{propWithDefault.Name.ToSnakeCase()}: {(propWithDefault.Type.IsNullable ? "Optional[" : string.Empty)}{returnType}{(propWithDefault.Type.IsNullable ? "]" : string.Empty)} = {propWithDefault.DefaultValue}");
+            writer.WriteLine();
+        }
+    }
+    private void WriteSetterAccessProperties(CodeClass parentClass, LanguageWriter writer) {
+        foreach(var propWithDefault in parentClass.GetPropertiesOfKind(SetterAccessProperties)
+                                        .Where(static x => !string.IsNullOrEmpty(x.DefaultValue))
+                                        .OrderByDescending(static x => x.Kind)
+                                        .ThenBy(static x => x.Name)) {                                
+            writer.WriteLine($"self.{propWithDefault.Name.ToSnakeCase()} = {propWithDefault.DefaultValue}");
+        }
+    }
+    private void WriteSetterAccessPropertiesWithoutDefaults(CodeClass parentClass, LanguageWriter writer) {
+        foreach(var propWithoutDefault in parentClass.GetPropertiesOfKind(SetterAccessProperties)
+                                        .Where(static x => string.IsNullOrEmpty(x.DefaultValue))
+                                        .OrderByDescending(static x => x.Kind)
+                                        .ThenBy(static x => x.Name)) {
+            var returnType = conventions.GetTypeString(propWithoutDefault.Type, propWithoutDefault, true, writer);
+            conventions.WriteInLineDescription(propWithoutDefault.Description, writer);
+            writer.WriteLine($"self.{conventions.GetAccessModifier(propWithoutDefault.Access)}{propWithoutDefault.NamePrefix}{propWithoutDefault.Name.ToSnakeCase()}: {(propWithoutDefault.Type.IsNullable ? "Optional[" : string.Empty)}{returnType}{(propWithoutDefault.Type.IsNullable ? "]" : string.Empty)} = None");
         }
     }
     private static void AssignPropertyFromParameter(CodeClass parentClass, CodeMethod currentMethod, CodeParameterKind parameterKind, CodePropertyKind propertyKind, LanguageWriter writer, string variableName = default) {