+++ tests/Kiota.Builder.Tests/Writers/Python/CodeMethodWriterTests.cs
@@ -313,13 +313,6 @@ public class CodeMethodWriterTests : IDisposable {
     }
     [Fact]
     public void WritesDeSerializerBody() {
-        var parameter = new CodeParameter{
-            Description = ParamDescription,
-            Name = ParamName
-        };
-        parameter.Type = new CodeType {
-            Name = "string"
-        };
         method.Kind = CodeMethodKind.Deserializer;
         method.IsAsync = false;
         AddSerializationProperties();
@@ -334,6 +327,7 @@ public class CodeMethodWriterTests : IDisposable {
         Assert.Contains("get_collection_of_primitive_values", result);
         Assert.Contains("get_collection_of_object_values", result);
         Assert.Contains("get_enum_value", result);
+        Assert.DoesNotContain("defined_in_parent", result, StringComparison.OrdinalIgnoreCase);
     }
     [Fact]
     public void WritesInheritedSerializerBody() {
@@ -347,13 +341,6 @@ public class CodeMethodWriterTests : IDisposable {
     }
     [Fact]
     public void WritesSerializerBody() {
-        var parameter = new CodeParameter{
-            Description = ParamDescription,
-            Name = ParamName
-        };
-        parameter.Type = new CodeType {
-            Name = "string"
-        };
         method.Kind = CodeMethodKind.Serializer;
         method.IsAsync = false;
         AddSerializationProperties();
@@ -364,6 +351,7 @@ public class CodeMethodWriterTests : IDisposable {
         Assert.Contains("write_collection_of_object_values", result);
         Assert.Contains("write_enum_value", result);
         Assert.Contains("write_additional_data_value(self.additional_data)", result);
+        Assert.DoesNotContain("defined_in_parent", result, StringComparison.OrdinalIgnoreCase);
     }
     [Fact]
     public void WritesMethodAsyncDescription() {
@@ -411,7 +399,7 @@ public class CodeMethodWriterTests : IDisposable {
     }
     [Fact]
     public void Defensive() {
-        var codeMethodWriter = new CodeMethodWriter(new PythonConventionService());
+        var codeMethodWriter = new CodeMethodWriter(new PythonConventionService(), false);
         Assert.Throws<ArgumentNullException>(() => codeMethodWriter.WriteCodeElement(null, writer));
         Assert.Throws<ArgumentNullException>(() => codeMethodWriter.WriteCodeElement(method, null));
         var originalParent = method.Parent;
@@ -615,6 +603,49 @@ public class CodeMethodWriterTests : IDisposable {
         Assert.Contains("has a description", result);
         Assert.Contains($"self.{propName}: Optional[str] = {defaultValue}", result);
         Assert.Contains("get_path_parameters", result);
+    }
+    [Fact]
+    public void WritesConstructorWithInheritance() {
+        method.Kind = CodeMethodKind.Constructor;
+        method.IsAsync = false;
+        var propName = "prop_with_no_default_value";
+        parentClass.Kind = CodeClassKind.Model;
+        AddInheritanceClass();
+        parentClass.AddProperty(new CodeProperty {
+            Name = propName,
+            Kind = CodePropertyKind.Custom,
+            Description = "This property has a description",
+            Type = new CodeType {
+                Name = "string"
+            }
+        });
+        var defaultValue = "someVal";
+        var prop2Name = "prop_with_default_value";
+        parentClass.Kind = CodeClassKind.RequestBuilder;
+        parentClass.AddProperty(new CodeProperty {
+            Name = prop2Name,
+            DefaultValue = defaultValue,
+            Kind = CodePropertyKind.UrlTemplate,
+            Description = "This property has a description",
+            Type = new CodeType {
+                Name = "string"
+            }
+        });
+        AddRequestProperties();
+        method.AddParameter(new CodeParameter {
+            Name = "pathParameters",
+            Kind = CodeParameterKind.PathParameters,
+            Type = new CodeType {
+                Name = "Union[Dict[str, Any], str]",
+                IsNullable = true,
+            }
+        });
+        writer.Write(method);
+        var result = tw.ToString();
+        Assert.Contains("super().__init__()", result);
+        Assert.Contains("has a description", result);
+        Assert.Contains($"self.{prop2Name}: Optional[str] = {defaultValue}", result);
+        Assert.Contains($"self.{propName}: Optional[str] = None", result);
     }
     [Fact]
     public void WritesApiConstructor() {